+++ src/main/java/com/ll/townforest/boundedContext/home/controller/AdminController.java
@@ -1,5 +1,9 @@
 package com.ll.townforest.boundedContext.home.controller;
 
+import java.util.List;
+import java.util.stream.Collectors;
+
+import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Slice;
 import org.springframework.security.access.prepost.PreAuthorize;
@@ -14,6 +18,8 @@ import org.springframework.web.bind.annotation.ResponseBody;
 import com.ll.townforest.base.rq.Rq;
 import com.ll.townforest.base.rsData.RsData;
 import com.ll.townforest.boundedContext.apt.entity.AptAccount;
+import com.ll.townforest.boundedContext.gym.entity.GymMembership;
+import com.ll.townforest.boundedContext.gym.service.GymService;
 import com.ll.townforest.boundedContext.library.entity.LibraryHistory;
 import com.ll.townforest.boundedContext.library.entity.Seat;
 import com.ll.townforest.boundedContext.library.service.LibraryService;
@@ -26,6 +32,8 @@ import lombok.RequiredArgsConstructor;
 public class AdminController {
 	private final Rq rq;
 	private final LibraryService libraryService;
+
+	private final GymService gymService;
 
 	@GetMapping("")
 	@PreAuthorize("isAuthenticated()")
@@ -74,5 +82,39 @@ public class AdminController {
 		AptAccount targetUser = libraryService.getTargetUserForAdminCancel(libraryHistoryId);
 
 		return libraryService.adminCancel(targetUser, canCancelSeat.getData(), libraryHistoryId).getMsg();
+	}
+
+	@PreAuthorize("isAuthenticated()")
+	@GetMapping("/gym")
+	public String adminMain(Model model) {
+		AptAccount user = rq.getAptAccount();
+		if (!rq.isGymAdmin())
+			rq.historyBack("헬스장 관리자만 접속 가능합니다.");
+
+		List<GymMembership> currentUsers = gymService.getMemberList(user);
+		model.addAttribute("currentUsers", currentUsers);
+
+		// 이용권 정지시킨 회원 필터링
+		List<GymMembership> pauseUsers = currentUsers
+			.stream()
+			.filter(a -> a.getStatus() == 3)
+			.collect(Collectors.toList());
+
+		model.addAttribute("pauseUsers", pauseUsers);
+
+		return "admin/gym/gym";
+	}
+
+	@PreAuthorize("isAuthenticated()")
+	@GetMapping("/gym/member")
+	public String showMember(Model model, @RequestParam(value = "page", defaultValue = "0") int page) {
+		AptAccount user = rq.getAptAccount();
+		if (!rq.isGymAdmin())
+			rq.historyBack("헬스장 관리자만 접속 가능합니다.");
+
+		Page<GymMembership> paging = gymService.getMemberPage(page, user);
+		model.addAttribute("paging", paging);
+
+		return "admin/gym/members";
 	}
 }
